Improve app error handling and refactor plugin
