+++ PetIA-app-bridge/composer.json
@@ -1,0 +1,23 @@
+{
+  "name": "labotica/petia-app-bridge",
+  "type": "wordpress-plugin",
+  "require": {
+    "firebase/php-jwt": "^6.10"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^11.1"
+  },
+  "autoload": {
+    "psr-4": {
+      "": "includes/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "PetIAAppBridge\\Tests\\": "tests/"
+    }
+  },
+  "scripts": {
+    "test": "phpunit"
+  }
+}
