+++ PetIA-app-bridge/phpunit.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<phpunit bootstrap="vendor/autoload.php" colors="false">
+  <testsuites>
+    <testsuite name="Plugin Tests">
+      <directory>tests</directory>
+    </testsuite>
+  </testsuites>
+</phpunit>
