+++ PetIA/js/app.js
@@ -1,6 +1,7 @@
 import config from '../config.js';
 import { logout, validateToken } from './auth.js';
 import { fetchWithAuth, getToken } from './token.js';
+import { handleError } from './error.js';
 
 async function getProfile() {
   const valid = await validateToken();
@@ -75,7 +76,7 @@ async function updateProfile(e) {
     alert('Datos actualizados');
     getProfile();
   } else {
-    alert('Error al actualizar');
+    handleError(new Error('Error al actualizar'));
   }
 }
