+++ PetIA/js/error.js
@@ -1,0 +1,6 @@
+export function handleError(err, userMessage = 'Ocurrió un error') {
+  console.error(err);
+  if (typeof window !== 'undefined' && typeof window.alert === 'function') {
+    alert(userMessage);
+  }
+}
